Backend/requirements.txt
PyJWT
pydantic
pydantic-settings
fastapi
uvicorn
sqlalchemy
pandas
python-jose[cryptography]
python-multipart
openpyxl
email-validator
mysqlclient
mysql-connector-python
pymysql
cryptography
